Move sequential LC derivation out of matching
