+++ src/colmap/controllers/option_manager_test.cc
@@ -140,7 +140,6 @@ TEST(OptionManager, WriteAndRead) {
   options_write.feature_extraction->sift->max_num_features = 4096;
   options_write.sequential_pairing->mark_loop_detection_as_lc = true;
   options_write.sequential_pairing->mark_non_consecutive_as_lc = true;
-  options_write.vocab_tree_pairing->mark_matches_as_lc = true;
   options_write.mapper->min_num_matches = 20;
   options_write.global_mapper->mapper.track_lc_second_pass = true;
   options_write.global_mapper->mapper.global_positioning.use_lc_observations =
@@ -177,8 +176,6 @@ TEST(OptionManager, WriteAndRead) {
             options_write.sequential_pairing->mark_loop_detection_as_lc);
   EXPECT_EQ(options_read.sequential_pairing->mark_non_consecutive_as_lc,
             options_write.sequential_pairing->mark_non_consecutive_as_lc);
-  EXPECT_EQ(options_read.vocab_tree_pairing->mark_matches_as_lc,
-            options_write.vocab_tree_pairing->mark_matches_as_lc);
   EXPECT_EQ(options_read.mapper->min_num_matches,
             options_write.mapper->min_num_matches);
   EXPECT_EQ(options_read.global_mapper->mapper.track_lc_second_pass,
