+++ src/colmap/controllers/pairing.h
@@ -37,7 +37,6 @@
 
 #include <filesystem>
 #include <unordered_set>
-#include <vector>
 
 namespace colmap {
 
@@ -75,10 +74,6 @@ struct VocabTreePairingOptions {
 
   // Optional path to file with specific image names to match.
   std::filesystem::path match_list_path = "";
-
-  // Whether verified pairs generated by this matcher should be marked as loop
-  // closures in the database.
-  bool mark_matches_as_lc = false;
 
   // Number of threads for indexing and retrieval.
   int num_threads = -1;
@@ -153,11 +148,12 @@ struct SequentialPairingOptions {
   int loop_detection_max_num_features = -1;
 
   // Whether verified loop-detection pairs should be marked as loop closures in
-  // the database.
+  // the database by the sequential LC postprocess.
   bool mark_loop_detection_as_lc = false;
 
   // Whether non-consecutive sequential-overlap pairs should be marked as loop
-  // closure candidates. Consecutive pairs remain tracking/non-LC seed pairs.
+  // closure candidates by the sequential LC postprocess. Consecutive pairs
+  // remain tracking/non-LC seed pairs.
   bool mark_non_consecutive_as_lc = false;
 
   // Number of threads for loop detection indexing and retrieval.
@@ -243,22 +239,6 @@ struct ExistingMatchedPairingOptions {
   }
 };
 
-struct FeatureMatcherImagePair {
-  image_t image_id1 = kInvalidImageId;
-  image_t image_id2 = kInvalidImageId;
-  // Pair-level provenance used by feature matching and verification. Sequential
-  // distance-1 tracking pairs keep this false; loop-detection and extended
-  // sequential candidates may set it true.
-  bool is_loop_closure = false;
-};
-
-std::vector<FeatureMatcherImagePair> FeatureMatcherImagePairs(
-    const std::vector<std::pair<image_t, image_t>>& image_pairs,
-    bool mark_as_loop_closure = false);
-
-std::vector<std::pair<image_t, image_t>> ImagePairsWithoutProvenance(
-    const std::vector<FeatureMatcherImagePair>& image_pairs);
-
 class PairGenerator {
  public:
   virtual ~PairGenerator() = default;
@@ -268,8 +248,6 @@ class PairGenerator {
   virtual bool HasFinished() const = 0;
 
   virtual std::vector<std::pair<image_t, image_t>> Next() = 0;
-
-  virtual std::vector<FeatureMatcherImagePair> NextWithProvenance();
 
   std::vector<std::pair<image_t, image_t>> AllPairs();
 };
@@ -318,8 +296,6 @@ class VocabTreePairGenerator : public PairGenerator {
 
   std::vector<std::pair<image_t, image_t>> Next() override;
 
-  std::vector<FeatureMatcherImagePair> NextWithProvenance() override;
-
  private:
   void IndexImages(const std::vector<image_t>& image_ids);
 
@@ -358,11 +334,8 @@ class SequentialPairGenerator : public PairGenerator {
 
   std::vector<std::pair<image_t, image_t>> Next() override;
 
-  std::vector<FeatureMatcherImagePair> NextWithProvenance() override;
-
  private:
   std::vector<image_t> GetOrderedImageIds() const;
-  std::vector<std::pair<image_t, image_t>> NextSequentialPairs();
 
   const SequentialPairingOptions options_;
   const std::shared_ptr<FeatureMatcherCache> cache_;
