+++ src/colmap/controllers/pairing_test.cc
@@ -129,26 +129,14 @@ TEST(VocabTreePairGenerator, Nominal) {
   }
 }
 
-TEST(VocabTreePairingOptions, LoopClosureProvenanceFlag) {
-  VocabTreePairingOptions options;
-  EXPECT_FALSE(options.mark_matches_as_lc);
-
-  options.mark_matches_as_lc = true;
-  EXPECT_TRUE(options.mark_matches_as_lc);
-}
-
-TEST(SequentialPairingOptions, VocabTreeOptionsLoopClosureProvenance) {
+TEST(SequentialPairingOptions, LoopClosurePostprocessFlags) {
   SequentialPairingOptions options;
-  options.mark_loop_detection_as_lc = true;
-
-  EXPECT_TRUE(options.VocabTreeOptions().mark_matches_as_lc);
-}
-
-TEST(SequentialPairingOptions, NonConsecutiveLoopClosureProvenanceFlag) {
-  SequentialPairingOptions options;
+  EXPECT_FALSE(options.mark_loop_detection_as_lc);
   EXPECT_FALSE(options.mark_non_consecutive_as_lc);
 
+  options.mark_loop_detection_as_lc = true;
   options.mark_non_consecutive_as_lc = true;
+  EXPECT_TRUE(options.mark_loop_detection_as_lc);
   EXPECT_TRUE(options.mark_non_consecutive_as_lc);
 }
 
@@ -178,85 +166,10 @@ TEST(SequentialPairGenerator, Linear) {
                   std::make_pair(images[2].ImageId(), images[3].ImageId()),
                   std::make_pair(images[2].ImageId(), images[4].ImageId())));
   EXPECT_THAT(generator.Next(),
-              testing::ElementsAre(std::make_pair(images[3].ImageId(),
-                                                  images[4].ImageId())));
+              testing::ElementsAre(
+                  std::make_pair(images[3].ImageId(), images[4].ImageId())));
   EXPECT_TRUE(generator.Next().empty());
   EXPECT_TRUE(generator.HasFinished());
-}
-
-TEST(SequentialPairGenerator, LinearDefaultUsesVanillaPerImageBatching) {
-  constexpr int kNumImages = 5;
-  auto database = Database::Open(kInMemorySqliteDatabasePath);
-  CreateSyntheticDatabase(kNumImages, *database);
-  const std::vector<Image> images = database->ReadAllImages();
-  CHECK_EQ(images.size(), kNumImages);
-
-  SequentialPairingOptions options;
-  options.overlap = 3;
-  options.quadratic_overlap = false;
-  SequentialPairGenerator generator(options, database);
-
-  EXPECT_THAT(generator.Next(),
-              testing::ElementsAre(
-                  std::make_pair(images[0].ImageId(), images[1].ImageId()),
-                  std::make_pair(images[0].ImageId(), images[2].ImageId()),
-                  std::make_pair(images[0].ImageId(), images[3].ImageId())));
-  const auto pairs = generator.NextWithProvenance();
-  EXPECT_THAT(ImagePairsWithoutProvenance(pairs),
-              testing::ElementsAre(
-                  std::make_pair(images[1].ImageId(), images[2].ImageId()),
-                  std::make_pair(images[1].ImageId(), images[3].ImageId()),
-                  std::make_pair(images[1].ImageId(), images[4].ImageId())));
-  for (const FeatureMatcherImagePair& pair : pairs) {
-    EXPECT_FALSE(pair.is_loop_closure);
-  }
-}
-
-TEST(SequentialPairGenerator, LoopClosureFlagKeepsVanillaPairBatches) {
-  constexpr int kNumImages = 5;
-  auto database = Database::Open(kInMemorySqliteDatabasePath);
-  CreateSyntheticDatabase(kNumImages, *database);
-  const std::vector<Image> images = database->ReadAllImages();
-  CHECK_EQ(images.size(), kNumImages);
-
-  SequentialPairingOptions options;
-  options.overlap = 3;
-  options.quadratic_overlap = false;
-  options.mark_non_consecutive_as_lc = true;
-  SequentialPairGenerator generator(options, database);
-
-  const auto pairs = generator.NextWithProvenance();
-  EXPECT_THAT(ImagePairsWithoutProvenance(pairs),
-              testing::ElementsAre(
-                  std::make_pair(images[0].ImageId(), images[1].ImageId()),
-                  std::make_pair(images[0].ImageId(), images[2].ImageId()),
-                  std::make_pair(images[0].ImageId(), images[3].ImageId())));
-  for (const FeatureMatcherImagePair& pair : pairs) {
-    EXPECT_FALSE(pair.is_loop_closure);
-  }
-}
-
-TEST(SequentialPairGenerator, LinearOverlapProvenanceDefaultIsVanilla) {
-  constexpr int kNumImages = 5;
-  auto database = Database::Open(kInMemorySqliteDatabasePath);
-  CreateSyntheticDatabase(kNumImages, *database);
-  const std::vector<Image> images = database->ReadAllImages();
-  CHECK_EQ(images.size(), kNumImages);
-
-  SequentialPairingOptions options;
-  options.overlap = 3;
-  options.quadratic_overlap = false;
-  SequentialPairGenerator generator(options, database);
-
-  const auto pairs = generator.NextWithProvenance();
-  EXPECT_THAT(ImagePairsWithoutProvenance(pairs),
-              testing::ElementsAre(
-                  std::make_pair(images[0].ImageId(), images[1].ImageId()),
-                  std::make_pair(images[0].ImageId(), images[2].ImageId()),
-                  std::make_pair(images[0].ImageId(), images[3].ImageId())));
-  for (const FeatureMatcherImagePair& pair : pairs) {
-    EXPECT_FALSE(pair.is_loop_closure);
-  }
 }
 
 TEST(SequentialPairGenerator, LinearRig) {
@@ -319,7 +232,6 @@ TEST(SequentialPairGenerator, Quadratic) {
   SequentialPairingOptions options;
   options.overlap = 3;
   options.quadratic_overlap = true;
-  options.mark_non_consecutive_as_lc = true;
   SequentialPairGenerator generator(options, database);
   EXPECT_THAT(generator.Next(),
               testing::ElementsAre(
@@ -335,8 +247,8 @@ TEST(SequentialPairGenerator, Quadratic) {
                   std::make_pair(images[2].ImageId(), images[3].ImageId()),
                   std::make_pair(images[2].ImageId(), images[4].ImageId())));
   EXPECT_THAT(generator.Next(),
-              testing::ElementsAre(std::make_pair(images[3].ImageId(),
-                                                  images[4].ImageId())));
+              testing::ElementsAre(
+                  std::make_pair(images[3].ImageId(), images[4].ImageId())));
   EXPECT_TRUE(generator.Next().empty());
   EXPECT_TRUE(generator.HasFinished());
 }
