+++ src/colmap/controllers/sequential_loop_closure.h
@@ -1,0 +1,59 @@
+// Copyright (c), ETH Zurich and UNC Chapel Hill.
+// All rights reserved.
+//
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions are met:
+//
+//     * Redistributions of source code must retain the above copyright
+//       notice, this list of conditions and the following disclaimer.
+//
+//     * Redistributions in binary form must reproduce the above copyright
+//       notice, this list of conditions and the following disclaimer in the
+//       documentation and/or other materials provided with the distribution.
+//
+//     * Neither the name of ETH Zurich and UNC Chapel Hill nor the names of
+//       its contributors may be used to endorse or promote products derived
+//       from this software without specific prior written permission.
+//
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+// ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDERS OR CONTRIBUTORS BE
+// LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+// CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+// SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+// INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+// CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+// ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+// POSSIBILITY OF SUCH DAMAGE.
+
+#pragma once
+
+#include "colmap/controllers/matcher_cache.h"
+#include "colmap/controllers/pairing.h"
+
+#include <filesystem>
+#include <functional>
+#include <memory>
+
+namespace colmap {
+
+bool SequentialLoopClosurePostprocessEnabled(
+    const SequentialPairingOptions& options);
+
+// Derive LC provenance from the verified sequential-matcher database state.
+// Direct consecutive image pairs and same/adjacent-frame rig pairs are kept as
+// non-LC tracking pairs. Enabled non-consecutive sequential-overlap and/or
+// loop-detection pairs are compared against transitive direct-track evidence;
+// transitive rows stay non-LC and remaining candidate rows become LC.
+void DeriveSequentialLoopClosureProvenance(
+    const std::shared_ptr<FeatureMatcherCache>& cache,
+    const SequentialPairingOptions& options,
+    const std::function<bool()>& is_stopped = nullptr);
+
+void DeriveSequentialLoopClosureProvenance(
+    const std::filesystem::path& database_path,
+    const SequentialPairingOptions& options,
+    const std::function<bool()>& is_stopped = nullptr);
+
+}  // namespace colmap
