+++ src/pycolmap/pipeline/match_features.cc
@@ -1,5 +1,6 @@
 #include "colmap/controllers/feature_matching.h"
 #include "colmap/controllers/pairing.h"
+#include "colmap/controllers/sequential_loop_closure.h"
 #include "colmap/estimators/two_view_geometry.h"
 #include "colmap/exe/feature.h"
 #include "colmap/exe/sfm.h"
@@ -14,6 +15,7 @@
 
 #include <filesystem>
 #include <memory>
+#include <type_traits>
 
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
@@ -49,6 +51,9 @@ void MatchFeatures(const std::filesystem::path& database_path,
       pairing_options, matching_options, verification_options, database_path);
   matcher->Start();
   PyWait(matcher.get());
+  if constexpr (std::is_same_v<PairingOptions, SequentialPairingOptions>) {
+    DeriveSequentialLoopClosureProvenance(database_path, pairing_options);
+  }
 }
 
 void VerifyMatches(const std::filesystem::path& database_path,
@@ -161,11 +166,6 @@ void BindMatchFeatures(py::module& m) {
               "match_list_path",
               &VocabTreePairingOptions::match_list_path,
               "Optional path to file with specific image names to match.")
-          .def_readwrite(
-              "mark_matches_as_lc",
-              &VocabTreePairingOptions::mark_matches_as_lc,
-              "Whether verified pairs generated by this matcher should be "
-              "marked as loop closures in the database.")
           .def_readwrite("num_threads", &VocabTreePairingOptions::num_threads)
           .def("check", &VocabTreePairingOptions::Check);
   MakeDataclass(PyVocabTreePairingOptions);
